refactor: optimize ReportFormattedName and reuse DisplayedName in Banner
