+++ packages/nameguard-react/src/components/Report/Banner.tsx
@@ -1,13 +1,14 @@
 import React from "react";
 import cc from "classcat";
 import { type NameGuardReport } from "@namehash/nameguard";
-import type { ParsedName } from "@namehash/ens-utils";
+import { type ParsedName } from "@namehash/ens-utils";
 
 import { RatingIcon, RatingIconSize } from "./RatingIcon";
 import { RatedBox } from "../RatedBox/RatedBox";
 import { ReportChangesApplied } from "./ReportChangesApplied";
-import { ReportFormattedDisplayName } from "./ReportFormattedName";
+import { ReportFormattedDisplayName } from "./ReportFormattedDisplayName";
 import { ratingTextColor } from "../../utils/text";
+import { DisplayedName } from "../DisplayedName/DisplayedName";
 
 type Props = {
   report?: NameGuardReport;
@@ -15,8 +16,7 @@ type Props = {
 };
 
 export function Banner({ report, parsedName }: Props) {
-  const { name, title, subtitle, rating, beautiful_name, normalization } =
-    report;
+  const { title, subtitle, rating } = report;
 
   const text = cc([
     "font-semibold text-sm md:text-2xl",
@@ -30,14 +30,15 @@ export function Banner({ report, parsedName }: Props) {
           <p className="uppercase text-[12px] text-gray-500 font-medium">
             Rating for
           </p>
-          <h1 className="pt-1 text-2xl md:text-4xl text-black font-semibold md:font-bold overflow-hidden overflow-ellipsis whitespace-nowrap ens-webfont">
-            {name}
+          <h1>
+            <DisplayedName
+              textStylingClasses="pt-1 text-2xl md:text-4xl text-black font-semibold md:font-bold overflow-hidden overflow-ellipsis whitespace-nowrap ens-webfont"
+              name={parsedName.outputName}
+              maxDisplayWidth={600}
+              displayRawName={true}
+            />
           </h1>
-          <ReportFormattedDisplayName
-            beautifiedName={beautiful_name}
-            normalization={normalization}
-            name={parsedName.outputName}
-          />
+          <ReportFormattedDisplayName name={parsedName.outputName} />
         </div>
         <div className="flex items-start space-x-4 pt-5 md:pt-0 md:w-2/6 flex-shrink-0">
           <div className="flex-shrink-0">
